feat(auth): integrate Auth0 for real authorization, removing dummyJSON login functionality
