+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@auth0/nextjs-auth0": "^3.5.0",
     "@fortawesome/fontawesome-free": "^6.6.0",
     "js-cookie": "^3.0.5",
     "lodash": "^4.17.21",
