+++ src/app/(dashboard)/blogs/[blogId]/page.js
