+++ src/app/(dashboard)/blogs/blogs.css
