+++ src/app/(dashboard)/products/[productId]/product-details.css
