+++ src/app/api/auth/[auth0]/route.js
@@ -1,0 +1,3 @@
+import { handleAuth } from "@auth0/nextjs-auth0";
+
+export const GET = handleAuth();
