--- src/app/components/AuthChecker.jsx
@@ -1,29 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
-import Cookies from "js-cookie";
-
-import Spinner from "../components/Spinner/Spinner";
-
-export default function AuthChecker({ children }) {
-  const router = useRouter();
-
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const token = Cookies.get("accessToken");
-
-    if (!token) {
-      router.push("/login");
-    } else {
-      setLoading(false);
-    }
-  }, [router]);
-
-  if (loading) {
-    return <Spinner message="Loading..." />;
-  }
-
-  return <>{children}</>;
-}
