--- src/app/components/LogOutBtn/LogOutBtn.css
@@ -1,18 +1,0 @@
-.btn-logout {
-  cursor: pointer;
-  background-color: var(--color-border);
-  padding: 0.8rem 1.2rem;
-  border-radius: var(--radius-md);
-  border: none;
-  color: var(--color-text);
-  font-size: 1.6rem;
-  transform: translateY(0);
-  transition: all 0.2s;
-}
-
-.btn-logout:hover {
-  transform: translateY(-1px);
-}
-.btn-logout:active {
-  transform: translateY(1px);
-}
