+++ src/app/layout.js
@@ -1,3 +1,4 @@
+import { UserProvider } from "@auth0/nextjs-auth0/client";
 import "./globals.css";
 
 export const metadata = {
@@ -9,7 +10,7 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body>
-        <div id="root">{children}</div>
+        <UserProvider>{children}</UserProvider>
       </body>
     </html>
   );
